+++ src/Credfeto.Package.Push/Configuration/ServiceConfiguration.cs
@@ -5,12 +5,13 @@ using ILogger = NuGet.Common.ILogger;
 
 namespace Credfeto.Package.Push.Configuration;
 
-internal static class ServiceConfiguration
+public static class ServiceConfiguration
 {
     public static IServiceCollection AddServices(this IServiceCollection services)
     {
         return services
             .AddSingleton<ILogger, NugetForwardingLogger>()
+            .AddSingleton<IPackagePushGateway, NuGetPackagePushGateway>()
             .AddSingleton<IUploadOrchestration, UploadOrchestration>()
             .AddSingleton<IPackageUploader, PackageUploader>();
     }